Move requestLocalStoragePermission() code to main activity
